Grab all active FSN's no matter which module they are specified in and include concepts with modified descriptions/definitions.
The US edition sometimes deactivates an international FSN and supplies their own.  All concepts need a FSN, so (not entirely correctly) we use whichever FSN is active regardless of module. The US edition also changes / adds descriptions to concepts defined in the CORE module, so we need to reflect the changes when emitting the US edition.  test/conf/en_us_stated.json has a config to generate the US edition and test/data/FSNChange has a US edition with one concept (407309009) for testing purposes.